requirements.txt
pandas>=2.0
gspread>=5.0.0
gspread-dataframe>=4.0.0
google-auth>=2.0.0
apscheduler>=3.9.0
python-dotenv>=1.0.0
openpyxl>=3.0.0